A curtain of green
Got the pattern for the Graze enemy down, but there's a weird bug in the
main level where, from the side, a curtain of green just starts falling
down as it cruises past the screen. Another thing is that, on the main
level, the pattern does not start correctly, nor does it start for all
the copies of the enemy.