Half-fix
The bug itself on the life assignment is corrected, but a new bug arose
(likely due to the messy code you need to clean) that has the player
spawn before checking the amount of lives. Furthermore, whenever the
player dies, the number of lives gets assigned to the default value.
Relevant scripts: Character_Player and PlayerSpawn